FS-4286: reusing number of calls to get fund and found round details
